% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cc_regression.R
\name{run_cc_regression}
\alias{run_cc_regression}
\title{Cell Cycle Scoring and SCTransform Regression}
\usage{
run_cc_regression(obj, species = "mouse")
}
\arguments{
\item{obj}{A Seurat object.}

\item{species}{Character string. Currently supports "mouse" (default) and "human".
It automatically handles gene capitalization (Title Case for mouse, All Caps for human).}
}
\value{
A Seurat object with a new 'SCT' assay containing regressed data.
}
\description{
This function performs cell cycle scoring based on species-specific gene lists
and uses SCTransform to regress out cell cycle scores ("S.Score" and "G2M.Score").
}
